+++ .changeset/lovely-tigers-breathe.md
@@ -1,0 +1,5 @@
+---
+'@signalwire/realtime-api': patch
+---
+
+Fix `onStarted` function in decorated promises
