+++ internal/e2e-realtime-api/src/voiceDetect/withAllListeners.test.ts
@@ -6,7 +6,7 @@ import {
   createTestRunner,
   TestHandler,
   makeSipDomainAppAddress,
-} from './utils'
+} from '../utils'
 
 const handler: TestHandler = ({ domainApp }) => {
   if (!domainApp) {
