+++ internal/e2e-realtime-api/src/voicePlayback/withDialListeners.test.ts
@@ -6,7 +6,7 @@ import {
   CALL_PROPS,
   TestHandler,
   makeSipDomainAppAddress,
-} from './utils'
+} from '../utils'
 
 const handler: TestHandler = ({ domainApp }) => {
   if (!domainApp) {
@@ -25,7 +25,7 @@ const handler: TestHandler = ({ domainApp }) => {
       })
 
       const unsubVoice = await client.voice.listen({
-        topics: [domainApp.call_relay_context],
+        topics: [domainApp.call_relay_context, 'home'],
         onCallReceived: async (call) => {
           try {
             const resultAnswer = await call.answer()
@@ -51,48 +51,45 @@ const handler: TestHandler = ({ domainApp }) => {
           domain: domainApp.domain,
         }),
         timeout: 30,
+        listen: {
+          onStateChanged: async (call) => {
+            if (call.state === 'ended') {
+              await unsubVoice()
+
+              await client.disconnect()
+
+              resolve(0)
+            }
+          },
+          onPlaybackStarted: (playback) => {
+            tap.hasProps(playback, CALL_PLAYBACK_PROPS, 'Playback started')
+            tap.equal(playback.state, 'playing', 'Playback correct state')
+          },
+          onPlaybackUpdated: (playback) => {
+            tap.notOk(playback.id, 'Playback updated')
+          },
+          onPlaybackFailed: (playback) => {
+            tap.notOk(playback.id, 'Playback failed')
+          },
+          onPlaybackEnded: async (playback) => {
+            tap.hasProps(playback, CALL_PLAYBACK_PROPS, 'Playback ended')
+            tap.equal(playback.state, 'finished', 'Playback correct state')
+
+            await call.hangup()
+          },
+        },
       })
       tap.ok(call.id, 'Outbound - Call resolved')
 
-      const unsubCall = await call.listen({
-        onStateChanged: async (call) => {
-          if (call.state === 'ended') {
-            await unsubVoice()
+      const play = await call
+        .playAudio({
+          url: 'https://cdn.signalwire.com/default-music/welcome.mp3',
+        })
+        .onStarted()
 
-            await unsubCall?.()
-
-            await client.disconnect()
-
-            resolve(0)
-          }
-        },
-        onPlaybackStarted: (playback) => {
-          tap.hasProps(playback, CALL_PLAYBACK_PROPS, 'Playback started')
-          tap.equal(playback.state, 'playing', 'Playback correct state')
-        },
-        onPlaybackUpdated: (playback) => {
-          tap.notOk(playback.id, 'Playback updated')
-        },
-        onPlaybackFailed: (playback) => {
-          tap.notOk(playback.id, 'Playback failed')
-        },
-        onPlaybackEnded: async (playback) => {
-          tap.hasProps(playback, CALL_PLAYBACK_PROPS, 'Playback ended')
-          tap.equal(playback.state, 'finished', 'Playback correct state')
-
-          await call.hangup()
-        },
-      })
-
-      const play = call.playAudio({
-        url: 'https://cdn.signalwire.com/default-music/welcome.mp3',
-      })
-
-      if ((await play.state) === 'playing') {
-        await play.stop()
-      }
+      await play.stop()
     } catch (error) {
-      console.error('VoicePlaybackCallListeners error', error)
+      console.error('VoicePlaybackDialListeners error', error)
       reject(4)
     }
   })
@@ -100,7 +97,7 @@ const handler: TestHandler = ({ domainApp }) => {
 
 async function main() {
   const runner = createTestRunner({
-    name: 'Voice Playback with Call Listeners E2E',
+    name: 'Voice Playback with Dial Listeners E2E',
     testHandler: handler,
     executionTime: 60_000,
     useDomainApp: true,
